// features/api/model/organizationsIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { OrganizationResource } from './organizationResource';
import type { OrganizationsIndex200Links } from './organizationsIndex200Links';
import type { OrganizationsIndex200Meta } from './organizationsIndex200Meta';

export type OrganizationsIndex200 = {
  data: OrganizationResource[];
  links: OrganizationsIndex200Links;
  meta: OrganizationsIndex200Meta;
};